[IMP] sale: pricelist rule clarification message
Before this commit:
As discount policy is removed from pricelist it is becoming quite hard for user
to tell when discount is applied and when it's not.

After this commit:
While selecting price type in pricelist rules user will able to see when
original price is displayed on sale order line and website and when it is not.

Add some alert boxes for user information based on price type selected, when
discounts feature in enabled from settings.

affected version - master
task - 4502296
