+++ addons/sale/__manifest__.py
@@ -47,6 +47,7 @@ This module contains all the common features of Sales Management and eCommerce.
         'views/mail_activity_plan_views.xml',
         'views/payment_views.xml',
         'views/product_document_views.xml',
+        'views/product_pricelist_item_views.xml',
         'views/product_template_views.xml',
         'views/product_views.xml',
         'views/res_partner_views.xml',
