+++ addons/sale/views/product_pricelist_item_views.xml
@@ -1,0 +1,49 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="product_pricelist_item_form" model="ir.ui.view">
+        <field name="name">product.pricelist.item.view.form.inherit</field>
+        <field name="model">product.pricelist.item</field>
+        <field name="inherit_id" ref="product.product_pricelist_item_form_view"/>
+        <field name="arch" type="xml">
+            <group name="pricelist_rule_limits" position="replace">
+                <div>
+                    <group name="pricelist_rule_limits">
+                        <field name="min_quantity" string="Min Qty"/>
+                        <field
+                            name="date_start"
+                            string="Validity Period"
+                            widget="daterange"
+                            options="{'end_date_field': 'date_end'}"
+                        />
+                        <field name="date_end" invisible="1"/>
+                    </group>
+                    <div groups="sale.group_discount_per_so_line">
+                        <div
+                            class="alert alert-info"
+                            role="alert"
+                            invisible="compute_price != 'percentage'"
+                        >
+                            <div id="discount_price_warning">
+                                In sale order line original price is unit price and discount is in
+                                discount column.
+                            </div>
+                        </div>
+                        <div
+                            class="alert alert-danger"
+                            role="alert"
+                            invisible="compute_price == 'percentage'"
+                        >
+                            <div id="formula_fixed_price_warning">
+                                When prices are calculated through formula or fixed, original price is NOT
+                                displayed in sale order lines. Calculated/fixed price is set as unit price
+                                with 0% discount.
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </group>
+        </field>
+    </record>
+
+</odoo>
